% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_format.R
\name{fpca_format}
\alias{fpca_format}
\title{Format functional data ready for FPCA analysis}
\usage{
fpca_format(data)
}
\arguments{
\item{data}{original data}
}
\value{
an fpca formatted dataset
}
\description{
Format functional data ready for FPCA analysis
}
