% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregMFPCA.R
\name{irregMFPCA}
\alias{irregMFPCA}
\title{Function to carry out irregular MFPCA}
\usage{
irregMFPCA(components = NULL, split = F, ...)
}
\arguments{
\item{components}{restrict number of components to use}

\item{split}{whether to split the output}

\item{...}{list of fpca results}
}
\value{
eigenfunctions and vectors from mfpca
}
\description{
Function to carry out irregular MFPCA
}
