% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMFPCA.R
\name{simMFPCA}
\alias{simMFPCA}
\title{Function to simulate irregular functional data}
\usage{
simMFPCA(
  seed = 16,
  D = matrix(c(3, 0, 0, 1), nrow = 2, ncol = 2, byrow = T),
  N = 73,
  c = 2,
  t = seq(0, 1, length.out = 51),
  n_obs = 10
)
}
\arguments{
\item{seed}{random seed for replication}

\item{D}{variance matrix}

\item{N}{number of subjects}

\item{c}{number of components}

\item{t}{time vector; by default, a sequence from 0 to 1 with length 51}

\item{n_obs}{number of non-missing observations; must be shorter than t}
}
\value{
entirely simulated data
}
\description{
Function to simulate irregular functional data
}
