% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble_format.R
\name{tibble_format}
\alias{tibble_format}
\title{# Function that formats data into a tibble}
\usage{
tibble_format(dat, tindex = seq(0, 1, length.out = 51))
}
\arguments{
\item{dat}{original data}

\item{tindex}{time vector}
}
\value{
a tibble formatted data
}
\description{
# Function that formats data into a tibble
}
