% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_format.R
\name{vector_filler}
\alias{vector_filler}
\title{Helper function to fill vector from a tibble}
\usage{
vector_filler(v, target, source)
}
\arguments{
\item{v}{vector}

\item{target}{target to fill}

\item{source}{source to fill from}
}
\value{
a filled vector
}
\description{
Helper function to fill vector from a tibble
}
